Move StageMenu to AsciiManager
